interview-bot/backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
openai
whisperx
python-dotenv
torch
torchaudio
python-multipart